Cancel job when finding a 404 page.
When requesting a chapter page, Mangatown often displays a "404 Page not found" page. When that happens, the job was simply ended and it resulted in an empty folder being created for the chapter.

We now send back an object with a "cancel" code, which should tell manganese that this chapter will not actually be downloaded and the folder should be removed.